requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
alembic==1.13.3
pydantic==2.9.2
pydantic[email]==2.9.2
email-validator==2.2.0
pydantic-settings==2.5.2
jinja2==3.1.4
python-multipart==0.0.12
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
itsdangerous==2.2.0
slowapi==0.1.9
fastapi-csrf-protect==1.0.7
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
